analisis_riesgo_crediticio_v2/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# AWS and Cloud Storage
boto3>=1.26.0
botocore>=1.29.0

# Data Science and ML
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.2.0

# Data Processing
pyarrow>=10.0.0  # For Parquet support
openpyxl>=3.0.0  # For Excel support

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0

# Model Versioning and Experiment Tracking
mlflow>=2.0.0
dvc[s3]>=3.55.0  # Data Version Control with S3 support

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code Quality
black>=22.0.0
isort>=5.10.0

# Monitoring and Logging
loguru>=0.6.0

# Configuration Management
pydantic>=1.10.0
pyyaml>=6.0
